requirements.txt
flwr==1.22.0
flwr-datasets>=0.3.0
torch==2.9.0
torchvision==0.24.0
numpy>=1.21.0
absl-py>=1.0.0